bwd_split_kernel: migrate loading of k^T off block pointers.
No performance loss for case fused-attention-batch512-head32-d64-bwd-causal=False
