+++ tritonsrc/bwd_kernel_common.py
@@ -16,7 +16,7 @@ def dot(BLOCK_M : tl.constexpr, QDIM : tl.constexpr, KDIM : tl.constexpr, q, k):
 
 @triton.jit
 def bwd_kernel_dk_dv_common(
-    Q_block_ptr, KT_block_ptr, VT_block_ptr, B_block_ptr,
+    Q_block_ptr, kt, VT_block_ptr, B_block_ptr,
     sm_scale, DO_block_ptr,
     DK_block_ptr, DV_block_ptr,
     l_ptrs,
@@ -49,14 +49,7 @@ def bwd_kernel_dk_dv_common(
     qk_scale = sm_scale * 1.44269504089
     # load k and v: they will stay in SRAM throughout
     # (BLOCK_DMODEL, BLOCK_N)
-    '''
-    if PADDED_HEAD:
-        kt = tl.load(KT_block_ptr, boundary_check=(1,0), padding_option="zero")
-    else:
-        kt = tl.load(KT_block_ptr, boundary_check=(1,), padding_option="zero")
-    '''
-    kt = tl.load(KT_block_ptr)
-    kt = (kt * qk_scale).to(KT_block_ptr.type.element_ty)
+    kt = (kt * qk_scale).to(kt.type.element_ty)
     # (BLOCK_DMODEL, BLOCK_N)
     '''
     if PADDED_HEAD:
@@ -216,11 +209,15 @@ def bwd_kernel_dq_db_common(
     offs_n = tl.arange(0, BLOCK_N)
     # Check for OOB accesses on D and LSE
     overflow_size_q = start_m + BLOCK_M - seqlen_q
-    boundary = tl.full((BLOCK_M, ), BLOCK_M - overflow_size_q, dtype=tl.int32)
-    d_lse_ptrs_mask = boundary > tl.arange(0, BLOCK_M)
-    d_lse_padding = tl.full((BLOCK_M, ), 0, dtype=tl.float32)
-    Di = tl.load(D_ptrs + offs_m, mask=d_lse_ptrs_mask, other=d_lse_padding)
-    l_i = tl.load(l_ptrs + offs_m, mask=d_lse_ptrs_mask, other=d_lse_padding)
+    if overflow_size_q > 0:
+        boundary = tl.full((BLOCK_M, ), BLOCK_M - overflow_size_q, dtype=tl.int32)
+        d_lse_ptrs_mask = boundary > tl.arange(0, BLOCK_M)
+        d_lse_padding = tl.full((BLOCK_M, ), 0, dtype=tl.float32)
+        Di = tl.load(D_ptrs + offs_m, mask=d_lse_ptrs_mask, other=d_lse_padding)
+        l_i = tl.load(l_ptrs + offs_m, mask=d_lse_ptrs_mask, other=d_lse_padding)
+    else:
+        Di = tl.load(D_ptrs + offs_m)
+        l_i = tl.load(l_ptrs + offs_m)
     dq = tl.zeros([BLOCK_M, BLOCK_DMODEL], dtype=tl.float32)
     # loop over k, v
     lo = 0
